remove Flask from train_data_service module
Includes some house keeping.
